Moved all Alias/Catalog methods from Util to new AliasUtil class
